Separate router actions/loaders from components for fast refresh
- Extract event action from EventForm.jsx to actions/eventActions.js
- Extract course action from CourseForm.jsx to actions/courseActions.js
- Move combinedLoader, checkAuthLoader, eventDetailLoader to loaders/loaders.js
- Move logoutAction from auth.js to authActions.js (group all actions)
- Remove loaders/actions from auth.js (now purely auth concerns)
- Remove eventDetailLoader from util/util.js (now purely utility functions)
- Delete unused Admin.jsx (duplicate of Login with syntax errors)
- Update App.jsx imports to use new locations

Components now only export components, fixing React Fast Refresh.
