+++ src/components/EventForm.jsx
@@ -1,12 +1,9 @@
 import React, { useState } from "react";
-import { getAuthToken } from "../auth/auth";
-
 import {
   Form,
   useNavigate,
   useNavigation,
   useActionData,
-  redirect,
 } from "react-router-dom";
 
 const EventForm = ({ method, event = {} }) => {
@@ -472,80 +469,3 @@ const EventForm = ({ method, event = {} }) => {
 };
 
 export default EventForm;
-
-export async function action({ request, params }) {
-  const method = request.method;
-  const data = await request.formData();
-
-  const eventData = {
-    title: data.get("title"),
-    shortDescription: data.get("shortDescription"),
-    longDescription: data.get("description"),
-    date: data.get("date"),
-    openingTime: data.get("openingTime"),
-    street: data.get("street"),
-    postCode: data.get("postCode"),
-    city: data.get("city"),
-    ageRestriction: data.get("ageRestriction"),
-    accessibilityInfo: data.get("accessibilityInfo"),
-    ticketPrice: parseFloat(data.get("ticketPrice")),
-    featured: data.get("featured") === "true",
-    isReoccurring: data.get("isReoccurring") === "true",
-    reoccurringStartDate: data.get("reoccurringStartDate") || null,
-    reoccurringEndDate: data.get("reoccurringEndDate") || null,
-    reoccurringFrequency: data.get("reoccurringFrequency") || null,
-    dayOfWeek: data.get("dayOfWeek") || null,
-    typeOfEvent: data.get("typeOfEvent") || "ASC",
-    isTournament: data.get("isTournament") === "true",
-    ticketsAvailable: data.get("ticketsAvailable") !== "" ? parseInt(data.get("ticketsAvailable"), 10) : undefined,
-  };
-
-  // Define token before using it
-  const token = getAuthToken();
-
-  const formData = new FormData();
-  formData.append("eventData", JSON.stringify(eventData));
-  const imageFile = data.get("image");
-  if (imageFile && imageFile instanceof File && imageFile.size > 0) {
-    formData.append("image", imageFile);
-  }
-
-  let url = `${import.meta.env.VITE_DEV_URI}events`;
-  if (method === "PUT" || method === "PATCH") {
-    const eventId = params.eventId;
-    url = `${import.meta.env.VITE_DEV_URI}events/${eventId}`;
-  }
-
-  try {
-    const response = await fetch(url, {
-      method: method,
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-      body: formData,
-    });
-
-    // Log response status
-
-    // Get response as text and try to parse as JSON
-    const responseText = await response.text();
-
-    let responseData;
-      responseData = JSON.parse(responseText);
-
-
-    if (!response.ok) {
-      return {
-        errors: {
-          message:
-            responseData.error || responseData.message || "Unknown error",
-        },
-      };
-    }
-
-    return redirect("/events");
-  } catch (error) {
-    console.error("Error during form submission:", error);
-    return { errors: { message: error.message } };
-  }
-}
